+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased-patch - ???
+
+### Fixed
+- Mastodon API: Search parameter `following` now correctly returns the followings rather than the followers
+
 ## [2.1.0] - 2020-08-28
 
 ### Changed